+++ src/main/java/sa/com/cloudsolutions/antikythera/parser/DepsolvingParser.java
@@ -8,11 +8,13 @@ import sa.com.cloudsolutions.antikythera.depsolver.DepSolver;
 import sa.com.cloudsolutions.antikythera.depsolver.Graph;
 import sa.com.cloudsolutions.antikythera.evaluator.ArgumentGenerator;
 import sa.com.cloudsolutions.antikythera.evaluator.NullArgumentGenerator;
+import sa.com.cloudsolutions.antikythera.evaluator.SpringEvaluator;
 
 import java.io.IOException;
 
 public abstract class DepsolvingParser {
     CompilationUnit cu;
+    protected SpringEvaluator evaluator;
 
     public void start() throws IOException {
         for(TypeDeclaration<?> decl : cu.getTypes()) {
@@ -52,6 +54,7 @@ public abstract class DepsolvingParser {
         }, null);
 
     }
+
 
     public abstract void evaluateMethod(MethodDeclaration md, ArgumentGenerator gen);
